Added random number generator.
The random number generator is a Linear congruential
psudeo random number generator that takes extra entropy
from changes in the Joypad state to increase its randomness.

I didn't add this to snesdev-common yet as I am unsure of this
generator type and would need to check to see if a linear
feedback would work just as well.